requirements.txt
dj-database-url==0.5.0
Django==2.1.5
django-heroku==0.3.1
Pillow==5.4.1
psycopg2==2.7.7
pytz==2018.9
whitenoise==4.1.2
django-taggit==0.23.0
Unidecode==1.0.23
py-trello
sentry-sdk==0.7.9
